refactor(itun): consolidate sheet-control scaffolding; decompose the god sheets
The sheet layer was where every new rules feature landed, and it had
grown the repo's worst duplication: TakeDamage/PilotTakeDamage ~200-line
near-clones, loadRef/numericTl copied verbatim, the role="alert"
advisory box hand-rolled at 12 sites, and the freshest-record idiom in
all 8 controls — around three files totalling 2,324 lines.

- New controlPrimitives.tsx (input classes, AdvisoryBox/Text,
  DamageIntakeRow + useDamageAmount, freshEntity) and refCatalog.ts;
  adopted across the control set.
- Sheet.tsx 891 -> 190: resolves entity/composition and dispatches to
  SheetPilotView/SheetMechView/SheetCrawlerView; the whole-store
  subscription + injectable-store test seam stay put, views receive
  resolved data via props (narrower re-render surfaces).
- PilotSheet 781 -> 394 (+PilotSheetItems), CrawlerSheet 652 -> 279
  (+CrawlerSheetItems).

Behavior-preserving: all 1220 ITUN tests pass with assertions untouched;
tsc + eslint --max-warnings 0 clean.

Audit sweep items 19+24/30.
